--- a/benchmarks/benchmark_dataset.py
+++ b/benchmarks/benchmark_dataset.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3 SPDX-License-Identifier: Apache-2.0
 # SPDX-License-Identifier: Apache-2.0
 """
 benchmark_dataset.py