--- a/benchmarks/benchmark_serving.py
+++ b/benchmarks/benchmark_serving.py
@@ -991,4 +991,4 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
     
-    (args)
+    main(args)