@JenZhao 2025-03-09
latest testing - checking why the sharegpt one looks different now 

**benchmark_serving.py main branch**

| Dataset | Backend | Successful requests | Benchmark duration (s) | Total input tokens |
|---------|---------|---------------------|--------------------------|--------------------|
| sonnet | openai-chat | 10         | 1.54       | 5409       |
| hf-vision-arena | openai-chat | 10         | 2.60       | 7191       |
| hf | openai-chat | 10         | 1.52       | 115        |
| sonnet | vllm | 10         | 1.53       | 5409       |
| sharegpt | vllm | 10         | 6.76       | 1374       |
| random | vllm | 10         | 1.49       | 10240      |
| burstgpt | vllm | 10         | 5.93       | 11970      |

**benchmark_serving.py latest change**

| Dataset | Backend | Successful requests | Benchmark duration (s) | Total input tokens |
|---------|---------|---------------------|--------------------------|--------------------|
| sonnet | openai-chat | 10         | 1.49       | 5409       |
| hf-vision-arena | openai-chat | 10         | 2.26       | 7191       |
| hf | openai-chat | 10         | 1.26       | 115        |
| sonnet | vllm | 10         | 1.50       | 5409       |
| sharegpt | vllm | 10         | 1.14       | 1960       |
| random | vllm | 10         | 1.45       | 10240      |
| burstgpt | vllm | 10         | 5.84       | 11970      |

**benchmark_throughput.py main branch**

main branch only has random and sharegpt dataset, its random dataset generation is different now
| Dataset | Processed Prompts | Throughput (requests/s) | Total tokens/s | Output tokens/s |
|---------|-------------------|-------------------------|----------------|-----------------|
| random | 10 | 6.83 | 7867.60 | 874.18 |
| ShareGPT_V3_unfiltered_cleaned_split.json | 10 | 1.60 | 725.71 | 345.22 |
| sonnet | 10 |  |  |  |
| burstgpt | 10 |  |  |  |

**benchmark_throughput.py latest change**

| Dataset | Processed Prompts | Throughput (requests/s) | Total tokens/s | Output tokens/s |
|---------|-------------------|-------------------------|----------------|-----------------|
| random | 10 | 6.33 | 7297.28 | 810.81 |
| ShareGPT_V3_unfiltered_cleaned_split.json | 10 | 3.25 | 1639.56 | 1152.57 |
| sonnet | 10 | 6.91 | 4557.63 | 1035.98 |
| burstgpt | 10 | 2.06 | 2853.16 | 386.93 |