@JenZhao 2025-03-09
testing again 

**Throughput Benchmark Results, this branch**

| Dataset | Processed Prompts | Throughput (requests/s) | Total tokens/s | Output tokens/s |
|---------|-------------------|-------------------------|----------------|-----------------|
| random | 10 | 6.80 | 7834.99 | 870.55 |
| ShareGPT_V3_unfiltered_cleaned_split.json | 10 | 3.44 | 883.61 | 434.41 |
| sonnet | 10 | 6.85 | 4496.04 | 1027.27 |
| burstgpt | 10 | 2.10 | 2906.03 | 394.10 |

**Throughput Benchmark Results, main branch**
* The main branch does not support Sonnet and BurstGPT.
* The main branch’s random dataset definition is different from that of this branch; this branch uses the random dataset definition in the serving script.
* This branch is also using serving script's sharegpt sampling. There is some very minor difference. 

| Dataset | Processed Prompts | Throughput (requests/s) | Total tokens/s | Output tokens/s |
|---------|-------------------|-------------------------|----------------|-----------------|
| random | 10 | 6.83 | 7867.60 | 874.18 |
| ShareGPT_V3_unfiltered_cleaned_split.json | 10 | 1.60 | 725.71 | 345.22 |
| sonnet | 10 |  |  |  |
| burstgpt | 10 |  |  |  |

----------

**Serving Benchmark Results, this branch**

| Dataset | Backend | Successful requests | Benchmark duration (s) | Total input tokens |
|---------|---------|---------------------|--------------------------|--------------------|
| sonnet | openai-chat | 10         | 1.53       | 5409       |
| hf-vision-arena | openai-chat | 10         | 2.61       | 7191       |
| hf | openai-chat | 10         | 1.53       | 115        |
| sonnet | vllm | 10         | 1.54       | 5409       |
| sharegpt | vllm | 10         | 6.53       | 1374       |
| random | vllm | 10         | 1.48       | 10240      |
| burstgpt | vllm | 10         | 5.92       | 11970      |

**Serving Benchmark Results, main branch**
| Dataset | Backend | Successful requests | Benchmark duration (s) | Total input tokens |
|---------|---------|---------------------|--------------------------|--------------------|
| sonnet | openai-chat | 10         | 1.54       | 5409       |
| hf-vision-arena | openai-chat | 10         | 2.60       | 7191       |
| hf | openai-chat | 10         | 1.52       | 115        |
| sonnet | vllm | 10         | 1.53       | 5409       |
| sharegpt | vllm | 10         | 6.76       | 1374       |
| random | vllm | 10         | 1.49       | 10240      |
| burstgpt | vllm | 10         | 5.93       | 11970      |

----------

**Lora Benchmark Results, this branch**

| Dataset | Num Prompts | Max Loras | Max Lora Rank | Enable Lora | Async Engine | Throughput (requests/s) | Total tokens/s | Output tokens/s |
|---------|-------------|-----------|---------------|-------------|--------------|-------------------------|----------------|-----------------|
| ShareGPT_V... | 10 | 1 | 8 | Yes | No | 1.72 | 705.30 | 327.45 |
| ShareGPT_V... | 10 | 4 | 8 | Yes | No | 1.34 | 663.50 | 230.17 |
| ShareGPT_V... | 10 | N/A | N/A | No | Yes | 1.94 | 1034.53 | 365.42 |
| ShareGPT_V... | 10 | 1 | 8 | Yes | Yes | 2.40 | 1014.93 | 379.04 |
| ShareGPT_V... | 10 | 4 | 8 | Yes | Yes | 0.99 | 633.56 | 359.78 |

**Lora Benchmark Results, main branch**

| Dataset | Num Prompts | Max Loras | Max Lora Rank | Enable Lora | Async Engine | Throughput (requests/s) | Total tokens/s | Output tokens/s |
|---------|-------------|-----------|---------------|-------------|--------------|-------------------------|----------------|-----------------|
| ShareGPT_V... | 10 | 1 | 8 | Yes | No | 1.65 | 819.24 | 467.69 |
| ShareGPT_V... | 10 | 4 | 8 | Yes | No | 1.41 | 736.43 | 240.00 |
| ShareGPT_V... | 10 | N/A | N/A | No | Yes | 1.77 | 773.02 | 375.51 |
| ShareGPT_V... | 10 | 1 | 8 | Yes | Yes | 1.85 | 736.69 | 443.24 |
| ShareGPT_V... | 10 | 4 | 8 | Yes | Yes | 1.76 | 729.75 | 240.61 |