@JenZhao 2025-03-09
testing with 1000 request

**Lora Benchmark Results, this branch**

| Dataset | Num Prompts | Max Loras | Max Lora Rank | Enable Lora | Async Engine | Throughput (requests/s) | Total tokens/s | Output tokens/s |
|---------|-------------|-----------|---------------|-------------|--------------|-------------------------|----------------|-----------------|
| ShareGPT_V... | 1000 | 1 | 8 | Yes | No | 12.05 | 5755.06 | 2816.05 |
| ShareGPT_V... | 1000 | 4 | 8 | Yes | No | 10.85 | 5351.16 | 2539.97 |
| ShareGPT_V... | 1000 | N/A | N/A | No | Yes | 16.52 | 8256.40 | 3949.34 |
| ShareGPT_V... | 1000 | 1 | 8 | Yes | Yes | 11.93 | 5582.34 | 2611.23 |
| ShareGPT_V... | 1000 | 4 | 8 | Yes | Yes | 11.21 | 5364.53 | 2653.42 |

**Lora Benchmark Results, main branch**

| Dataset | Num Prompts | Max Loras | Max Lora Rank | Enable Lora | Async Engine | Throughput (requests/s) | Total tokens/s | Output tokens/s |
|---------|-------------|-----------|---------------|-------------|--------------|-------------------------|----------------|-----------------|
| ShareGPT_V... | 1000 | 1 | 8 | Yes | No | 11.42 | 5434.65 | 2654.52 |
| ShareGPT_V... | 1000 | 4 | 8 | Yes | No | 10.87 | 5368.56 | 2520.88 |
| ShareGPT_V... | 1000 | N/A | N/A | No | Yes | 15.34 | 7528.82 | 3769.25 |
| ShareGPT_V... | 1000 | 1 | 8 | Yes | Yes | 11.60 | 5719.71 | 2573.52 |
| ShareGPT_V... | 1000 | 4 | 8 | Yes | Yes | 10.96 | 5139.18 | 2462.06 |